verify_system.py
#!/usr/bin/env python3
"""
DiatomsAI System Verification - Ensures everything is working beautifully
"""

import requests
import json
import sys
from datetime import datetime
from colorama import init, Fore, Style

init(autoreset=True)

def check_backend_health():
    """Verify backend is healthy and responsive"""
    try:
        response = requests.get('http://localhost:5002/api/health', timeout=5)
        if response.status_code == 200:
            data = response.json()
            health_score = data['database']['health_score']
            print(f"{Fore.GREEN}✅ Backend Health: {data['status']} (Score: {health_score})")
            return True
        else:
            print(f"{Fore.RED}❌ Backend unhealthy: Status {response.status_code}")
            return False
    except Exception as e:
        print(f"{Fore.RED}❌ Backend not accessible: {e}")
        return False

def check_frontend_status():
    """Verify React frontend is running"""
    try:
        response = requests.get('http://localhost:3000', timeout=5)
        if response.status_code == 200:
            print(f"{Fore.GREEN}✅ React Frontend: Running on port 3000")
            return True
        else:
            print(f"{Fore.RED}❌ Frontend issue: Status {response.status_code}")
            return False
    except Exception as e:
        print(f"{Fore.RED}❌ Frontend not accessible: {e}")
        return False

def check_analysis_quality():
    """Verify AI analysis quality and credibility"""
    try:
        response = requests.get('http://localhost:5002/api/analyses?limit=1')
        data = response.json()
        
        if data['analyses']:
            analysis = data['analyses'][0]
            quality_score = analysis.get('quality_score', 0)
            
            # Parse the raw JSON to check content quality
            if analysis.get('raw_response_json'):
                try:
                    content = json.loads(analysis['raw_response_json'])
                    
                    # Check for required high-quality fields
                    quality_checks = {
                        'Breaking News': bool(content.get('breaking_news')),
                        'Key Numbers': bool(content.get('key_numbers')),
                        'Market Pulse': bool(content.get('market_pulse')),
                        'Investment Ideas': bool(content.get('idea_desk')),
                        'Risk Analysis': bool(content.get('final_intel', {}).get('key_risks'))
                    }
                    
                    print(f"\n{Fore.CYAN}📊 Analysis Quality Check:")
                    for check, passed in quality_checks.items():
                        status = f"{Fore.GREEN}✓" if passed else f"{Fore.RED}✗"
                        print(f"  {status} {check}: {'Present' if passed else 'Missing'}")
                    
                    # Check for specific tickers/actionable advice
                    if content.get('idea_desk'):
                        print(f"\n{Fore.CYAN}💡 Investment Recommendations:")
                        for idea in content['idea_desk'][:3]:
                            action = idea.get('action', 'N/A')
                            ticker = idea.get('ticker', 'N/A')
                            print(f"  • {action} {ticker}")
                    
                    overall_quality = all(quality_checks.values())
                    if overall_quality and quality_score >= 3:
                        print(f"\n{Fore.GREEN}✅ Analysis Quality: EXCELLENT (Score: {quality_score}/10)")
                        return True
                    else:
                        print(f"\n{Fore.YELLOW}⚠️  Analysis Quality: NEEDS IMPROVEMENT (Score: {quality_score}/10)")
                        return False
                        
                except json.JSONDecodeError:
                    print(f"{Fore.RED}❌ Invalid analysis format")
                    return False
            else:
                print(f"{Fore.RED}❌ No analysis content found")
                return False
        else:
            print(f"{Fore.YELLOW}⚠️  No analyses available yet")
            return False
            
    except Exception as e:
        print(f"{Fore.RED}❌ Failed to check analysis: {e}")
        return False

def check_article_pipeline():
    """Verify article fetching and categorization"""
    try:
        response = requests.get('http://localhost:5002/api/stats')
        stats = response.json()['data']
        
        total_articles = stats.get('total_articles', 0)
        categories = stats.get('articles_by_category', {})
        
        print(f"\n{Fore.CYAN}📰 Article Pipeline Status:")
        print(f"  Total Articles: {total_articles}")
        
        if categories:
            print(f"  Categories:")
            for cat, count in categories.items():
                print(f"    • {cat.capitalize()}: {count}")
        
        if total_articles > 0:
            print(f"{Fore.GREEN}✅ Article Pipeline: Active")
            return True
        else:
            print(f"{Fore.YELLOW}⚠️  No articles in database")
            return False
            
    except Exception as e:
        print(f"{Fore.RED}❌ Failed to check articles: {e}")
        return False

def verify_credibility():
    """Verify the system is producing credible insights"""
    print(f"\n{Style.BRIGHT}{Fore.BLUE}🔍 CREDIBILITY VERIFICATION")
    print("=" * 50)
    
    # Check latest analysis for credibility markers
    try:
        response = requests.get('http://localhost:5002/api/articles?limit=5')
        articles = response.json()['data']
        
        if articles:
            print(f"\n{Fore.CYAN}📈 Recent Article Sources:")
            sources = set(article['source'] for article in articles)
            for source in list(sources)[:5]:
                print(f"  • {source}")
            
            # Check sentiment distribution
            sentiments = [article['sentiment_score'] for article in articles]
            avg_sentiment = sum(sentiments) / len(sentiments) if sentiments else 0
            
            print(f"\n{Fore.CYAN}💭 Sentiment Analysis:")
            print(f"  Average: {avg_sentiment:.2f}")
            print(f"  Range: [{min(sentiments):.2f}, {max(sentiments):.2f}]")
            
            return True
        else:
            print(f"{Fore.YELLOW}⚠️  No recent articles to verify")
            return False
            
    except Exception as e:
        print(f"{Fore.RED}❌ Credibility check failed: {e}")
        return False

def main():
    print(f"{Style.BRIGHT}{Fore.BLUE}🚀 DiatomsAI SYSTEM VERIFICATION")
    print(f"{Fore.BLUE}{'=' * 50}")
    print(f"Timestamp: {datetime.now().strftime('%Y-%m-%d %H:%M:%S')}\n")
    
    checks = [
        ("Backend API", check_backend_health),
        ("React Frontend", check_frontend_status),
        ("Article Pipeline", check_article_pipeline),
        ("AI Analysis Quality", check_analysis_quality),
        ("Credibility Check", verify_credibility)
    ]
    
    results = []
    for name, check_func in checks:
        print(f"\n{Fore.YELLOW}Checking {name}...")
        results.append(check_func())
    
    # Final verdict
    print(f"\n{Style.BRIGHT}{Fore.BLUE}{'=' * 50}")
    print(f"{Style.BRIGHT}{Fore.BLUE}FINAL VERDICT:")
    
    if all(results):
        print(f"{Style.BRIGHT}{Fore.GREEN}✅ SYSTEM IS WORKING BEAUTIFULLY!")
        print(f"{Fore.GREEN}✅ Producing credible, institutional-grade insights")
        print(f"{Fore.GREEN}✅ Ready for production use")
        
        print(f"\n{Fore.CYAN}Access Points:")
        print(f"  • Dashboard: http://localhost:3000")
        print(f"  • API: http://localhost:5002")
        print(f"  • Health: http://localhost:5002/api/health")
    else:
        failed_checks = [name for name, result in zip([c[0] for c in checks], results) if not result]
        print(f"{Style.BRIGHT}{Fore.RED}❌ ISSUES DETECTED:")
        for check in failed_checks:
            print(f"  • {check} needs attention")
        
        print(f"\n{Fore.YELLOW}Recommendations:")
        if "Backend API" in failed_checks:
            print("  1. Run: PORT=5002 python3 web_app.py")
        if "React Frontend" in failed_checks:
            print("  2. Run: cd frontend && npm start")
        if "AI Analysis Quality" in failed_checks:
            print("  3. Run bot to generate analysis: python3 main.py")

if __name__ == "__main__":
    main() 